+++ ChangeLog.rst
@@ -15,7 +15,7 @@ Scene Manager Version History
 * **New.04:** New Icon set
 * **New.04:** Show In Explorer Root/Raid folders added for image viewer
 * **Update.04:** Added project line to the project materials
-
+* **Update.05:** Search filter added to the Set Project Window
 
 2.1
 ---